+{s,g}et{res{g,u},u,g,eu}id()

+++ include/_exit/_exit.h
@@ -1,4 +1,4 @@
 #include "stdint.h"
 #include "stddef.h"
 
-extern void _exit(isize status) noreturn;
+extern void _exit(int status) noreturn;
